+++ src/core/markdown/adapters/type-to-renderable.ts
@@ -12,6 +12,7 @@ import {
   GetRenderableContentByTypeName,
   RenderableCustomObject,
   RenderableCustomField,
+  RenderableCustomMetadata,
 } from '../../renderables/types';
 import { adaptDescribable, adaptDocumentable } from '../../renderables/documentables';
 import { adaptConstructor, adaptMethod } from './methods-and-constructors';
@@ -21,6 +22,7 @@ import { ExternalMetadata, SourceFileMetadata } from '../../shared/types';
 import { CustomObjectMetadata } from '../../reflection/sobject/reflect-custom-object-sources';
 import { getTypeGroup, isInSource } from '../../shared/utils';
 import { CustomFieldMetadata } from '../../reflection/sobject/reflect-custom-field-source';
+import { CustomMetadataMetadata } from '../../reflection/sobject/reflect-custom-metadata-source';
 
 type GetReturnRenderable<T extends Type | CustomObjectMetadata> = T extends InterfaceMirror
   ? RenderableInterface
@@ -266,6 +268,12 @@ function objectMetadataToRenderable(
       heading: 'Fields',
       value: objectMetadata.fields.map((field) => fieldMetadataToRenderable(field, config, 3)),
     },
+    hasRecords: objectMetadata.metadataRecords.length > 0,
+    metadataRecords: {
+      headingLevel: 2,
+      heading: 'Records',
+      value: objectMetadata.metadataRecords.map((metadata) => customMetadataToRenderable(metadata, 3)),
+    },
   };
 }
 
@@ -289,6 +297,17 @@ function fieldMetadataToRenderable(
           value: field.pickListValues,
         }
       : undefined,
+  };
+}
+
+function customMetadataToRenderable(metadata: CustomMetadataMetadata, headingLevel: number): RenderableCustomMetadata {
+  return {
+    type: 'metadata',
+    headingLevel: headingLevel,
+    heading: metadata.label ?? metadata.name,
+    apiName: metadata.apiName,
+    label: metadata.label ?? metadata.name,
+    protected: metadata.protected,
   };
 }