+++ src/core/markdown/templates/custom-object-template.ts
@@ -39,4 +39,25 @@ export const customObjectTemplate = `
 {{/each}}
 {{/if}}
 
+{{#if hasRecords}}
+{{ heading metadataRecords.headingLevel metadataRecords.heading }}
+{{#each metadataRecords.value}}
+{{ heading headingLevel heading }}
+
+{{#if protected}}
+\`Protected\`
+{{/if}}
+
+{{#if description}}
+{{{renderContent description}}}
+{{/if}}
+
+**API Name**
+
+\`{{{apiName}}}\`
+
+{{#unless @last}}---{{/unless}}
+{{/each}}
+{{/if}}
+
 `.trim();