+++ src/util/source-bundle-utils.ts
@@ -1,6 +1,7 @@
 import {
   UnparsedApexBundle,
   UnparsedCustomFieldBundle,
+  UnparsedCustomMetadataBundle,
   UnparsedCustomObjectBundle,
   UnparsedSourceBundle,
 } from '../core/shared/types';
@@ -9,11 +10,11 @@ export function filterApexSourceFiles(sourceFiles: UnparsedSourceBundle[]): Unpa
   return sourceFiles.filter((sourceFile): sourceFile is UnparsedApexBundle => sourceFile.type === 'apex');
 }
 
-export function filterCustomObjectsAndFields(
+export function filterCustomObjectsFieldsAndMetadataRecords(
   sourceFiles: UnparsedSourceBundle[],
-): (UnparsedCustomObjectBundle | UnparsedCustomFieldBundle)[] {
+): (UnparsedCustomObjectBundle | UnparsedCustomFieldBundle | UnparsedCustomMetadataBundle)[] {
   return sourceFiles.filter(
     (sourceFile): sourceFile is UnparsedCustomObjectBundle =>
-      sourceFile.type === 'customobject' || sourceFile.type === 'customfield',
+      sourceFile.type === 'customobject' || sourceFile.type === 'customfield' || sourceFile.type === 'custommetadata',
   );
 }